% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_fix.R
\name{cc_fix}
\alias{cc_fix}
\title{ClusterCirc-fix}
\usage{
cc_fix(file, type = "scores", limits, p, m, w_com = "TRUE", w)
}
\arguments{
\item{file}{File to be used for the analysis.}

\item{type}{Is "scores" if the file contains raw scores of variables,
"loadings" if the file contains loadings from PCA or EFA.
Default is "scores".}

\item{limits}{Vector with the positions of the last item of each cluster,
e.g. c(6,10,18) for 3 clusters with C1 = i1-i6, C2 = i7-i10, C3 = i11-i18.}

\item{p}{Number of clusters.}

\item{m}{Number of variables.}

\item{w_com}{Is "TRUE" if communalities of the variables are used as weights.
Is "FALSE" if user-defined weights should be used. Default = "TRUE".}

\item{w}{Vector with weights for the variables. Weights need to be in the same
order as the variables in the data and should be the same weights as used in
cc_data for fair comparison of spc_w. Default = item communalities.}
}
\value{
cc_fix returns overall ClusterCirc coefficients for user-defined item
clusters: Overall ClusterCirc coefficients, coefficients for clusters and
for items. Results can be compared to results for item clustering as
suggested by cc_data.
}
\description{
Computes ClusterCirc coefficients for user-defined item
clusters. Items need to be sorted in the data file before analysis,
e.g. cluster 1 = items 1 to 6, cluster 2 = items 7 to 10, etc.
ClusterCirc-fix coefficients for user-defined item clusters can be compared
to ClusterCirc coefficients for item clusters found by ClusterCirc-Data.
}
\examples{
cc_fix(data_ex, type = "scores", limits = c(6,10,18), p = 3, m = 18, w_com = "TRUE", w)
}
