% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_loadings.R
\name{cc_loadings}
\alias{cc_loadings}
\title{ClusterCirc-Loadings}
\usage{
cc_loadings(A, m)
}
\arguments{
\item{A}{Loading matrix on two orthogonal axes (e.g. from PCA or EFA).}

\item{m}{Number of variables.}
}
\value{
Returns item angles to be inserted into ClusterCirc.
}
\description{
Procedure that extracts item angles by trigonometric transformation
of item loadings on two orthogonal axes.
}
