% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_raw.R
\name{cc_raw}
\alias{cc_raw}
\title{ClusterCirc: Raw algorithm}
\usage{
cc_raw(angles, comm, p, m, w, e, q)
}
\arguments{
\item{angles}{Item angles to perform ClusterCirc on.}

\item{comm}{Item communalities.}

\item{p}{Number of clusters.}

\item{m}{Number of variables.}

\item{w}{Vector with weights for the variables. Weights need to be in the same
order as the variables in the data. Default = item communalities.}

\item{e}{Cluster weight (0 <= e <= 1) defining the importance of within-cluster
proximity versus equal cluster spacing. Default is 1/p weighing all clusters
equally. e = 0: Maximum importance of between-cluster spacing, within-cluster
proximity is ignored. e = 1: Maximum importance of within-cluster proximity,
between-cluster spacing is ignored.}

\item{q}{Precision index for the algorithm. Precision is higher for larger
values. Default = 10.}
}
\value{
Returns item clusters with optimal circumplexity and ClusterCirc
coefficients: Overall ClusterCirc results, coefficients for clusters and for items.
}
\description{
Is a help function that performs the ClusterCirc search algorithm
within ClusterCirc-Data (cc_data) and ClusterCirc-Simu (cc_simu).
All of the arguments are taken from cc_data or cc_simu. cc_raw cannot be
performed on its own.
}
