% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circe_modfast.R
\name{circe_modfast}
\alias{circe_modfast}
\title{CircE-modified, fast}
\usage{
circe_modfast(
  R,
  v.names,
  m,
  N,
  r,
  equal.com,
  equal.ang,
  mcsc,
  start.values,
  ci.level = 0.95,
  factr = 1e+10,
  pgtol = 0,
  lmm = 3,
  iterlim = 100,
  upper = NULL,
  lower = NULL,
  print.level = 2,
  file = NULL,
  title = "Circumplex Estimation",
  try.refit.BFGS = FALSE,
  verbose = FALSE
)
}
\description{
This is a help function that performs Browne's procedure CIRCUM
on the items to obtain item angles and communalities to be inserted into
ClusterCirc. The function is a modified version of circe.bfgs from the CircE
package (Grassi et al., 2010)
}
