% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ex.R
\docType{data}
\name{data_ex}
\alias{data_ex}
\title{Exemplary data to test Cluster-Circ}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 300 rows and 18 columns.
}
\usage{
data_ex
}
\description{
Simulated data with perfect circumplexitiy in the population.
n_sample = 300 (sample size),
p = 3 (number of clusters),
n_var = 18 (number of variables, 6 variables per cluster).
}
\details{
To run Cluster-Circ on the exemplary data:
cc_data(file = data_ex, n_sample = 300, input = "PCA", p = 3,
n_var = 18, w_com = TRUE, w, comm, e_def = TRUE, e, q = 10)
}
\keyword{datasets}
