% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_data.R
\name{cc_data}
\alias{cc_data}
\title{Cluster-Circ Data}
\usage{
cc_data(file, type = "scores", p, m, q = 10)
}
\arguments{
\item{file}{File to be used for the analysis.}

\item{type}{Fill in "scores" if the file contains raw scores of variables,
"loadings" if the file contains loadings from PCA or EFA.
Default is "scores".}

\item{p}{Number of clusters.}

\item{m}{Number of variables.}

\item{q}{Precision index for the algorithm. Precision is higher for larger
values. Default = 10.}
}
\value{
Returns item clusters with optimal circumplexity and Cluster-Circ
indices: Overall Cluster-Circ results, indices for clusters and for items.
}
\description{
Finds item clusters with optimal circumplex spacing in your data.
If the file contains raw scores (type = "scores"), PCA without rotation
will be performed on the data before running Cluster-Circ.
}
\examples{
cc_data(file = data_ex, type = "scores", p = 3, m = 18, q = 10)

}
