% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_raw.R
\name{cc_raw}
\alias{cc_raw}
\title{Cluster-Circ: Raw algorithm}
\usage{
cc_raw(A, p, m, q)
}
\arguments{
\item{A}{Loading matrix to perform Cluster-Circ on.}

\item{p}{Number of clusters.}

\item{m}{Number of variables.}

\item{q}{Precision index for the algorithm. Precision is higher for larger
values. Default = 10. Must be an integer > 0.}
}
\value{
Returns item clusters with optimal circumplexity and Cluster-Circ
indices: Overall Cluster-Circ results, indices for clusters and for items.
}
\description{
Is a help function that performs the Cluster-Circ algorithm
within Cluster-Circ Data (cc_data) and Cluster-Circ Simu (cc_simu).
All of the arguments are taken from cc_data. cc_raw cannot be performed
on its own.
}
\examples{
cc_raw(A, p = 3, m = 18, q = 10)
}
