% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_simu.R
\name{cc_simu}
\alias{cc_simu}
\title{Cluster-Circ Simu}
\usage{
cc_simu(n)
}
\arguments{
\item{n}{= Number of subjects in the sample.}
}
\value{
Returns Cluster-Circ indices for the population data and results for
500 simulated samples (mean, SD, min, max). Empirical 'spacing with h²'
from your data will be compared to the mean value from the simulated
samples with perfect circumplexity to test whether circumplex fit of the
empirical data is acceptable (using the cumulative probability of the
standard normal distribution for p < .01, one-tailed).
}
\description{
Simulates data with perfect circumplex clusters in the population
using parameters from the empirical data (sample size, number of clusters,
number of items, empirical within-cluster range, mean item communality).
Performs Cluster-Circ on the simulated data for comparison with results
from Cluster-Circ Data. Number of simulated samples: 500.
}
\examples{
cc_simu(n = 100)
}
