% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ex.R
\docType{data}
\name{data_ex}
\alias{data_ex}
\title{Exemplary data to test Cluster-Circ}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 300 rows and 18 columns.
}
\usage{
data_ex
}
\description{
Simulated data with perfect circumplexitiy in the population.
N = 300 (sample size),
p = 3 (number of clusters),
m = 18 (number of variables, 6 variables per cluster).
}
\details{
To run Cluster-Circ on the exemplary data:
cc_data(file = data_ex, type = "scores", p = 3, m = 18, q = 10)
}
\keyword{datasets}
